Add files via upload
Project 1: Estimation of the sale price of vehicles 

All Algorithms could be valid since they converge except for the SGD aloghythm. The algorithm that best adapts to this use case is the random forest since, as it is not a linear regression, it can better adapt to the different data that enter it from the model. In addition, with randomForest we achieve better results than with Polynomial regression, since, although it manages to adapt much better to the model than a simple linear regression, it lacks the freedom that a random forest has to make different classifications in different areas: such as not predicting any negative value, or better adapt to outliers

Project 2: Unsupervised learning about network profiles

Uses of unsupervised learning to predict compatibility between profiles on a network.

Project 3: Is it going to rain tomorrow?

Uses of ML and DL to predict if it is going to rain the next day. It was the last and personal project which was submitted in the Kaggle competition finishing at  top 10 among 28 competitors.